Refactor import statements to use the updated visitors module structure
