+++ src/brms/instruments/personal_loan.py
@@ -6,7 +6,7 @@ from brms.instruments.base import Instrument
 from brms.instruments.registry import RetailInstrumentRegistry
 
 if TYPE_CHECKING:
-    from brms.instruments.visitor import Visitor
+    from brms.instruments.visitors import Visitor
 
 
 class PersonalLoan(Instrument):
